refactor: split quicksort.c and CLIcore_UI_history.c under 1000 lines
Phase 1 of file-size refactoring to enforce <1000 lines per C file.

quicksort.c (1202 lines) split into:
  - quicksort.c (338 lines): single-array sorts
  - quicksort_multiarray.c (884 lines): multi-array co-sort variants

CLIcore_UI_history.c (1428 lines) split into:
  - CLIcore_UI_history.c (271 lines): core history management
  - CLIcore_UI_history_display.c (878 lines): ghistory/lhistory display
  - CLIcore_UI_history_cmds.c (351 lines): expansion, search, save
